requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
pandas
pydantic
